requirements.txt
semantic-kernel
azure-identity
python-dotenv
